Improve RSID QSY functionality
The qsy() function will now change the rig frequency when it notices
that the modem is permanently frequency-locked (e.g., MT-63).

  * Remove modem_frequency argument and associated code from the various
    rig control set_qsy functions, this is now done by qsy().
  * Make modem::set_freq() virtual and override it in mt63 in order to
    better prevent frequency changes.
  * Use qsy() in the RSID decoder and notifier "Go to freq" callback.

+++ src/include/fl_digi.h
@@ -221,7 +221,7 @@ extern void setTabColors();
 
 extern void toggle_visible_modes(Fl_Widget*, void*);
 
-void qsy(long long rfc, long long fmid = -1LL);
+void qsy(long long rfc, int fmid = -1);
 
 void note_qrg(bool no_dup = true, const char* prefix = " ", const char* suffix = " ",
 	      trx_mode mode = NUM_MODES, long long rfc = 0LL, int afreq = 0);
